refactor generation of mock reference data
